Add files via upload
allows a pop up to be minized and expanded with the ingredients found. It is color coordinated for good = green , bad = red, and mid = yellow. If a bad ingredient is detected that will be the dominant pop up. This needs to be updated to have an icon show up when the popup is minimized.